split micrphone impl
