+++ libs/microphone/microphone.cpp
@@ -1,34 +1,8 @@
 #include "pxt.h"
-#include "dmac.h"
-#include "LevelDetector.h"
 #include "LevelDetectorSPL.h"
 
-#ifdef NRF52_SERIES
-#include "NRF52PDM.h"
-#define PDMDevice NRF52PDM
-#else
-#include "SAMDPDM.h"
-#define PDMDevice SAMD21PDM
-#endif
-
-#define MICROPHONE_MIN 52.0f
-#define MICROPHONE_MAX 120.0f
-
 namespace pxt {
-
-class WMicrophone {
-  public:
-    PDMDevice microphone;
-    LevelDetectorSPL level;
-    WMicrophone()
-        : microphone(*LOOKUP_PIN(MIC_DATA), *LOOKUP_PIN(MIC_CLOCK))
-        , level(microphone.output, 95.0, 75.0, 9, 52, DEVICE_ID_MICROPHONE)
-    {
-        microphone.enable();
-    }
-};
-SINGLETON(WMicrophone);
-
+    codal::LevelDetectorSPL* getMicrophoneLevel();
 }
 
 namespace input {
@@ -40,7 +14,7 @@ namespace input {
 //% parts="microphone"
 //% weight=88 blockGap=12
 void onLoudSound(Action handler) {
-    getWMicrophone(); // wake up service
+    getMicrophoneLevel(); // wake up service
     registerWithDal(DEVICE_ID_MICROPHONE, LEVEL_THRESHOLD_HIGH, handler);
 }
 
@@ -52,7 +26,10 @@ void onLoudSound(Action handler) {
 //% parts="microphone"
 //% weight=34 blockGap=8
 int soundLevel() {
-    const int micValue = getWMicrophone()->level.getValue();
+    auto level = getWMicrophoneLevel();
+    if (NULL == level)
+        return MICROPHONE_MIN;        
+    const int micValue = level->getValue();
     const int scaled = max(MICROPHONE_MIN, min(micValue, MICROPHONE_MAX)) - MICROPHONE_MIN;
     return min(0xff, scaled * 0xff / (MICROPHONE_MAX - MICROPHONE_MIN));
 }
@@ -66,8 +43,12 @@ int soundLevel() {
 //% value.min=1 value.max=255
 //% group="More" weight=14 blockGap=8
 void setLoudSoundThreshold(int value) {
+    auto level = getWMicrophoneLevel();
+    if (NULL == level)
+        return;
+
     value = max(0, min(0xff, value));
     const int scaled = MICROPHONE_MIN + value * (MICROPHONE_MAX - MICROPHONE_MIN) / 0xff;
-    getWMicrophone()->level.setHighThreshold(scaled);
+    level->setHighThreshold(scaled);
 }
 }
