+++ libs/microphone/pxt.json
@@ -4,6 +4,7 @@
     "files": [
         "README.md",
         "microphone.cpp",
+        "microphonehw.cpp",
         "jacdac.ts",
         "enums.d.ts",
         "shims.d.ts"
